refactor: restructure checkout screen and form components
