+++ src/screens/public/CheckoutScreen/CheckoutForm/CheckoutForm.tsx
@@ -5,19 +5,22 @@ import {
   useElements,
   PaymentElement,
 } from '@stripe/react-stripe-js';
-import { createStyles } from './styles';
 import { useColors } from '@theme/ThemeProvider';
+import { createStyles } from './styles';
 
 const CheckoutForm = () => {
   const stripe = useStripe();
   const elements = useElements();
   const [message, setMessage] = useState<string | null>(null);
   const [isLoading, setIsLoading] = useState(false);
+
   const colors = useColors();
   const styles = createStyles(colors);
 
   const handleSubmit = async () => {
-    if (!stripe || !elements) return;
+    if (!stripe || !elements) {
+      return;
+    }
 
     setIsLoading(true);
     setMessage(null);
@@ -35,7 +38,7 @@ const CheckoutForm = () => {
       if (error.type === 'card_error' || error.type === 'validation_error') {
         setMessage(error.message || 'Erro nos dados de pagamento.');
       } else {
-        setMessage('Ocorreu um erro inesperado.');
+        setMessage('Ocorreu um erro inesperado. Tente novamente.');
       }
       setIsLoading(false);
     } else {
@@ -44,18 +47,22 @@ const CheckoutForm = () => {
   };
 
   return (
-    <View>
-      <PaymentElement id="payment-element" />
+    <View style={styles.container}>
+      {/* Container para o Elemento Web do Stripe */}
+      <View style={styles.paymentElementContainer}>
+        <PaymentElement id="payment-element" />
+      </View>
 
       {message && <Text style={styles.errorMessageText}>{message}</Text>}
 
       <TouchableOpacity
         style={[
           styles.checkoutButton,
-          (isLoading || !stripe) && styles.checkoutButtonDisabled,
+          (isLoading || !stripe || !elements) && styles.checkoutButtonDisabled,
         ]}
         onPress={handleSubmit}
-        disabled={isLoading || !stripe || !elements}>
+        disabled={isLoading || !stripe || !elements}
+        activeOpacity={0.8}>
         {isLoading ? (
           <ActivityIndicator color={colors.primaryWhite} />
         ) : (
